+++ pylusat/utils.py
@@ -145,7 +145,67 @@ def read_raster(raster, nodata=None):
         raise ValueError("Not a valid raster data.")
 
 
-def random_ahp_weight(n=3):
+def _calc_ahp(r_mtx):
+    ri_dict = {3: 0.58, 4: 0.9, 5: 1.12, 6: 1.24,
+               7: 1.32, 8: 1.41, 9: 1.45, 10: 1.49}
+    n = r_mtx.shape[0]
+    eig_val, eig_vec = np.linalg.eig(r_mtx)
+    # principal eigenvalue, i.e., the highest valued eigenvalue
+    principal_eig_val = np.max(eig_val).real
+    principal_eig_vec = eig_vec[:, eig_val.argmax()].real
+
+    ci = (principal_eig_val - n) / (n - 1)   # consistency index
+    cr = ci / ri_dict[n]                     # consistency ratio
+
+    # the normalized principal eigenvector is called the priority vector
+    priority_vec = principal_eig_vec / np.sum(principal_eig_vec)
+    return priority_vec, cr
+
+
+def ahp(r_mtx):
+    """
+    Compute the priority vector of the Analytic Hierarchy Process (AHP).
+
+    The computed priority vector can be used as weights in a decision-making
+    process based on multiple criteria. The function also returns the
+    Consistency Ratio (CR) to check whether decisions of pair-wise comparisons
+    are consistent.
+
+    Parameters
+    ----------
+    r_mtx : np.ndarray
+        The reciprocal matrix.
+
+    Returns
+    -------
+    priority_vec : np.ndarray
+        The priority vector computed by AHP.
+    cr : float
+        Consistency Ratio (1990, Saaty). The rule of thumb is CR <= 0.1.
+    """
+    assert r_mtx.ndim == 2, "The reciprocal matrix must be 2-D."
+    nrow, ncol = r_mtx.shape
+    assert nrow == ncol, "The reciprocal matrix must be a square matrix."
+    # validate the number of criteria is between 3 and 10
+    assert 3 <= nrow <= 10, (
+        "The number of rows/columns must be between 3 and 10."
+    )
+    # validate matrix's elements are valid members of the AHP fundamental scale
+    ahp_scale = np.append(1 / (np.arange(9)+1), np.arange(1, 9)+1)
+    assert (
+        np.isin(r_mtx[np.tril_indices(nrow)], ahp_scale).all().item() is True
+    ), ("One or more elements of the input matrix "
+        "is not a valid member of "
+        "the AHP fundamental scale.")
+    # validate the matrix is indeed reciprocal
+    assert (
+        1 / r_mtx.T[np.tril_indices(nrow)] ==
+        r_mtx[np.tril_indices(nrow)]
+    ).all().item() is True, "Invalid reciprocal matrix."
+    return _calc_ahp(r_mtx)
+
+
+def random_ahp(n=3):
     """
     Generate randomized Analytic Hierarchy Process (AHP) weights.
 
@@ -160,36 +220,27 @@ def random_ahp_weight(n=3):
 
     Returns
     -------
-    np.ndarray
+    priority_vec : np.ndarray
         An array that contains the weights (priority vector of AHP).
     """
     if n < 3 or n > 10:
         raise ValueError("The number must be between 3 and 10.")
-    # dict of random consistency index
-    ri_dict = {3: 0.58, 4: 0.9, 5: 1.12, 6: 1.24,
-               7: 1.32, 8: 1.41, 9: 1.45, 10: 1.49}
 
     # an array of scales for pair-wise comparison
-    scale_pair_comp = np.append(1 / (np.arange(9) + 1), np.arange(1, 9) + 1)
+    ahp_scale = np.append(1 / (np.arange(9) + 1), np.arange(1, 9) + 1)
 
     while True:
         # upper triangular part of AHP's reciprocal matrix
-        triu_weights = np.random.choice(scale_pair_comp, np.sum(np.arange(n)))
+        triu_weights = np.random.choice(ahp_scale, np.sum(np.arange(n)))
 
-        ahp = np.identity(n)   # the reciprocal matrix
-        ahp[np.triu_indices(n, 1)] = triu_weights
-        ahp[np.tril_indices(n, -1)] = 1 / ahp.T[np.tril_indices(n, -1)]
+        r_mtx = np.identity(n)   # the reciprocal matrix
+        r_mtx[np.triu_indices(n, 1)] = triu_weights
+        r_mtx[np.tril_indices(n, -1)] = 1 / r_mtx.T[np.tril_indices(n, -1)]
 
-        eig_val, eig_vec = np.linalg.eig(ahp)
-        principal_eig_val = eig_val[0].real
-        principal_eig_vec = eig_vec[:, 0].real
-
-        ci = (principal_eig_val - n) / (n - 1)   # consistency index
-        cr = ci / ri_dict[n]                     # consistency ratio
+        priority_vec, cr = _calc_ahp(r_mtx)
         if cr < 0.1:       # general rule of thumb: CR less than 10%
             break
-    priority_vec = principal_eig_vec / np.sum(principal_eig_vec)
-    return priority_vec
+    return priority_vec, cr
 
 
 def weighted_sum(df, col_weights):
@@ -205,7 +256,7 @@ def weighted_sum(df, col_weights):
         column specified by the column name.
     Returns
     -------
-    np.ndarray
+    weighted_sum : np.ndarray
         An array contains the result of the weighted sum.
     """
     wgt_df = DataFrame([*col_weights.values()], index=[*col_weights])